Moved events to their own file and updated index.js accordingly
