+++ index.js
@@ -1,12 +1,9 @@
 require('dotenv').config()
 const fs = require('node:fs')
 const path = require('node:path')
-const { Client, Collection, GatewayIntentBits, MessageFlags } = require('discord.js')
+const { Client, Collection, GatewayIntentBits } = require('discord.js')
 const token = process.env.TOKEN
 const logger = require('./logger')
-const { hasCommandPermissionAsync } = require('./utils/guildSettings')
-const { connectToDatabase, setProfilingLevel, getSlowQueries } = require('./utils/database')
-const { runMigration } = require('./utils/migrateToMongoDB')
 
 logger.info('Starting DestroyerBot...')
 
@@ -65,162 +62,6 @@ for (const file of eventFiles) {
 }
 
 logger.info(`Loaded ${eventFiles.length} events.`)
-
-client.once('ready', async () => {
-  const { user, guilds, ws } = client
-
-  logger.info(
-    {
-      botUsername: user.username,
-      botDiscriminator: user.discriminator,
-      botId: user.id,
-      guildCount: guilds.cache.size,
-      websocketStatus: ws.status,
-      websocketPing: ws.ping,
-      readyAt: new Date().toISOString(),
-    },
-    `${user.tag} is online and ready! Serving ${guilds.cache.size} server(s)`
-  )
-
-  // Initialize database connection and run migration
-  try {
-    await connectToDatabase()
-    const migrated = await runMigration()
-    if (migrated) {
-      logger.info('Database migration completed successfully')
-    }
-
-    // Enable database performance monitoring in development
-    if (process.env.NODE_ENV !== 'production') {
-      // Enable mongoose debug logging for query monitoring
-      const mongoose = require('mongoose')
-      mongoose.set('debug', (collection, method, query, doc, options) => {
-        const start = Date.now()
-
-        // Log slow queries (>100ms) for performance monitoring
-        setImmediate(() => {
-          const duration = Date.now() - start
-          if (duration > 100) {
-            logger.warn(
-              {
-                collection,
-                method,
-                duration,
-                query: JSON.stringify(query).substring(0, 500), // Truncate long queries
-              },
-              'Slow database query detected'
-            )
-          }
-        })
-      })
-
-      // Try to enable MongoDB profiling for detailed query analysis
-      // Note: This may not be available on all MongoDB instances (e.g., Atlas)
-      try {
-        await setProfilingLevel(1) // Log queries slower than 100ms
-        logger.info('Database profiling enabled for performance monitoring')
-      } catch (profileError) {
-        logger.info(
-          {
-            error: profileError.message,
-            note: 'Database profiling not available (common with MongoDB Atlas or restricted instances). Query monitoring via mongoose debug logging is still active.',
-          },
-          'Database profiling unavailable - continuing without profiling'
-        )
-      }
-    }
-
-    // Periodic performance monitoring (every 30 minutes)
-    setInterval(
-      async () => {
-        try {
-          if (process.env.NODE_ENV !== 'production') {
-            const slowQueries = await getSlowQueries(5)
-            if (slowQueries.length > 0) {
-              logger.info(
-                { count: slowQueries.length, queries: slowQueries },
-                'Recent slow queries detected'
-              )
-            }
-          }
-        } catch (error) {
-          logger.debug({ error: error.message }, 'Performance monitoring check failed')
-        }
-      },
-      30 * 60 * 1000
-    ) // 30 minutes
-  } catch (error) {
-    logger.error({ error: error.message, stack: error.stack }, 'Failed to initialize database')
-    // Continue running even if database fails - will use JSON fallback
-  }
-})
-
-client.on('interactionCreate', async (interaction) => {
-  // Handle autocomplete interactions
-  if (interaction.isAutocomplete()) {
-    const command = client.commands.get(interaction.commandName)
-    if (!command || !command.autocomplete) return
-
-    try {
-      await command.autocomplete(interaction)
-    } catch (error) {
-      logger.error({ error: error.message, stack: error.stack }, 'Autocomplete error')
-    }
-    return
-  }
-
-  if (!interaction.isChatInputCommand()) return
-
-  const command = client.commands.get(interaction.commandName)
-
-  if (!command) return
-
-  // Check permissions for guild commands (skip for DMs)
-  if (interaction.guild && interaction.member) {
-    try {
-      const hasPermission = await hasCommandPermissionAsync(
-        interaction.guild.id,
-        interaction.commandName,
-        interaction.member
-      )
-
-      if (!hasPermission) {
-        logger.warn(
-          {
-            guildId: interaction.guild.id,
-            userId: interaction.user.id,
-            username: interaction.user.tag,
-            commandName: interaction.commandName,
-          },
-          'User blocked from using command due to permissions'
-        )
-        return interaction.reply({
-          content: "⛔ You don't have permission to use this command.",
-          flags: MessageFlags.Ephemeral,
-        })
-      }
-    } catch (error) {
-      logger.error(
-        { error: error.message, guildId: interaction.guild.id },
-        'Error checking command permissions'
-      )
-      return interaction.reply({
-        content: 'There was an error checking permissions. Please try again.',
-        flags: MessageFlags.Ephemeral,
-      })
-    }
-  }
-
-  try {
-    await command.execute(interaction)
-  } catch (error) {
-    logger.error(error)
-    await interaction.reply({
-      content: 'There was an error while executing this command!',
-      flags: MessageFlags.Ephemeral,
-    })
-  }
-})
 
 logger.info('Attempting to log in with the provided token...')
